Advanced Software Engineering Topic - Task 04 Reflection API / JUnit Testing Framework / Logging API Demonstration and Testing Result

+++ 6CS002_Task_04/src/com/wlv/reflection/reflection_5.java
@@ -19,7 +19,7 @@ public class reflection_5 {
     
     for(Field f : fields) {
       System.out.printf("field name=%s type=%s value=%d\n", 
-          f.getName(), f.getType(), f.getFloat(s));
+          f.getName(), f.getType(), f.getDouble(s));
       
     }
   }
