+++ 6CS002_Task_04/src/com/wlv/reflection/reflection_7.java
@@ -19,9 +19,10 @@ public class reflection_7 {
     for (Field f : fields) {
       f.setAccessible(true);
       System.out.printf("field name=%s type=%s value=%d\n", f.getName(),
-        f.getType(), f.getFloat(s));
+        f.getType(), f.getDouble(s));
     }
     
   }
 
 }
+
